requirements.txt
chameleon==2.15
coverage==3.7.1
docutils==0.11
mako==0.9.1
markupsafe==0.19
nose==1.3.1
passlib==1.6.1
pastedeploy==1.5.2
paster==0.8
pygments==1.6
pyramid==1.5
pyramid-chameleon==0.1
pyramid-debugtoolbar==2.0.2
pyramid-mako==1.0.1
pyramid-tm==0.7
repoze.lru==0.6
repoze.tm2==2.0
sqlalchemy==0.9.4
transaction==1.4.3
translationstring==1.1
venusian==1.0a8
waitress==0.8.8
webob==1.3.1
zope.deprecation==4.1.1
zope.interface==4.1.1
zope.sqlalchemy==0.7.4
boto==2.38.0
markerlib==0.4
sqlalchemy-searchable==0.5.0
sqlalchemy-utils==0.26.0
webhelpers==1.3.0
unidecode==0.04.16 
recaptcha==1.0rc1
python-slugify==0.0.7
pastescript==1.7.5
Pillow==2.5.0
requests==2.3.0